Add ui elements for choosing map seed

+++ src/main/java/guis/UpdateMapProperty.java
@@ -6,5 +6,5 @@ package guis;
  */
 public interface UpdateMapProperty {
 
-    void update(short value);
+    void update(long value);
 }
